requirements.txt
altgraph==0.17.4
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pefile==2023.2.7
psutil==5.9.8
py-cpuinfo==9.0.0
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
python-dateutil==2.8.2
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
six==1.16.0
tzdata==2024.1
WMI==1.5.1
XlsxWriter==3.1.9
